backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==5.0.3
django-cors-headers==4.5.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.1.0
psycopg2-binary==2.9.10
PyJWT==2.11.0
python-decouple==3.8
pytz==2025.2
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
